+++ src/game/entry_replacements.rs
@@ -589,6 +589,7 @@ impl Game {
             | AppliedEffectDef::CanBlockOnly(_)
             | AppliedEffectDef::PreventDamageFrom(_)
             | AppliedEffectDef::PreventCombatDamageFrom(_)
+            | AppliedEffectDef::RedirectPlayerDamageToThis(_)
             | AppliedEffectDef::PreventCombatDamage
             | AppliedEffectDef::PreventCombatDamageDealtBy
             | AppliedEffectDef::AddLandTypes(_)